Correcciones en las entidades y el repositorio

+++ app/src/main/java/com/swbvelasquez/nekoexpress/domain/usecase/GetProductByIdUseCase.kt
@@ -6,14 +6,16 @@ import com.swbvelasquez.nekoexpress.domain.model.ProductCatalogModel
 class GetProductByIdUseCase {
     private val repository = ProductRepository()
 
-    suspend operator fun invoke(id:Int):ProductCatalogModel?{
-        var product = repository.getProductByIdFromApi(id)
+    suspend operator fun invoke(productId:Long):ProductCatalogModel?{
+        var product = repository.getProductByIdFromDb(productId)
 
-        if(product!=null){
-            repository.deleteProductFromDb(product)
-            repository.insertProductToDb(product)
-        }else{
-            product = repository.getProductByIdFromDb(id)
+        if(product==null){
+            product = repository.getProductByIdFromApi(productId)
+
+            if(product!=null) {
+                repository.deleteProductFromDb(product)
+                repository.insertProductToDb(product)
+            }
         }
 
         return product
